Removed OperationPermissible from ComponentEntity.
