+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-client-dto/src/main/java/org/apache/nifi/web/api/entity/ComponentEntity.java
@@ -29,14 +29,13 @@ import java.util.Objects;
  * A base type for request/response entities.
  */
 @XmlRootElement(name = "componentEntity")
-public class ComponentEntity extends Entity implements OperationPermissible {
+public class ComponentEntity extends Entity {
 
     private RevisionDTO revision;
     private String id;
     private String uri;
     private PositionDTO position;
     private PermissionsDTO permissions;
-    private PermissionsDTO operatePermissions;
     private List<BulletinEntity> bulletins;
     private Boolean disconnectedNodeAcknowledged;
 
@@ -116,22 +115,6 @@ public class ComponentEntity extends Entity implements OperationPermissible {
 
     public void setPermissions(PermissionsDTO permissions) {
         this.permissions = permissions;
-    }
-
-    /**
-     * @return The permissions for this component operations
-     */
-    @ApiModelProperty(
-            value = "The permissions for this component operations."
-    )
-    @Override
-    public PermissionsDTO getOperatePermissions() {
-        return operatePermissions;
-    }
-
-    @Override
-    public void setOperatePermissions(PermissionsDTO permissions) {
-        this.operatePermissions = permissions;
     }
 
     /**
